refactor: 베이커리 form 구조 변경, input 컴포넌트 타입 수정
