--- src/components/BakeryDetail/index.ts
@@ -1,1 +1,0 @@
-export * from './Form';
